refactor(authcontext): replaced logic related to auth contex to a new file
replaced logic related to auth contex to a new file
